refactor: Add TokenImage
