+++ src/views/Farms/components/FarmCard/CardHeading.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { Tag, Flex, Heading } from '@pancakeswap/uikit'
 import { CommunityTag, CoreTag } from 'components/Tags'
 import { Token } from 'config/constants/types'
-import TokenPairImage from 'components/TokenPairImage'
+import { TokenPairImage } from 'components/TokenImage'
 
 export interface ExpandableSectionProps {
   lpLabel?: string
